rename web3 -> utils
