+++ packages/dapp/src/utils/connect.js
