+++ packages/dapp/src/utils/sign.js
